+++ lib/Community.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class CommunityScreen extends StatelessWidget {
+  const CommunityScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Community Screen'),
+    );
+  }
+}
