+++ lib/Market.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class MarketScreen extends StatelessWidget {
+  const MarketScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Market Screen'),
+    );
+  }
+}
