+++ lib/MyPage.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class MyPageScreen extends StatelessWidget {
+  const MyPageScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('My Page Screen'),
+    );
+  }
+}
